+++ tests-error/err_bad_literal4.in.c
@@ -1,0 +1,3 @@
+main() {
+	echo(0x1FFFFFFFF);
+}