+++ tests-error/err_undef_call.in.c
@@ -1,0 +1,3 @@
+main() {
+	this_doesnt_exist();
+}