+++ tests-error/err_undef_label.in.c
@@ -1,9 +1,5 @@
-#pragma renderer sds2
-
 main() {
-
 	if(1) {
 		goto baa;
 	}
-
 }