+++ tests-error/undef.in.c
@@ -1,8 +1,0 @@
-#pragma renderer sds2
-
-main() {
-
-
-	yolo = 12;
-
-}